agents: Add max_unsuccessful_requests, change how group change messages are handled
This commit stablishes a max unsuccessful request treshold
(`MAX_UNSUCCESSFUL_REQUESTS`), in order to follow the statement.

It also changes how the `TeacherGroupChangeMessage` is handled. It was
previously broadcasted to all the alumns, which made some alumns aware
of the change before the changed alumns. This was generating some wrong
denegation messages.

Now the teacher sends the message to the implicated alumns, and they
take care to broadcast the message when they have processed it.

This commit also introduces requests when changes are made, if some of
the new groups are interesting.